chap03/w2/src/main/java/org/fp024/w2/controller/LogoutController.java
package org.fp024.w2.controller;

import jakarta.servlet.ServletException;
import jakarta.servlet.annotation.WebServlet;
import jakarta.servlet.http.HttpServlet;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import jakarta.servlet.http.HttpSession;
import java.io.IOException;
import lombok.extern.slf4j.Slf4j;

@Slf4j
@WebServlet("/logout")
public class LogoutController extends HttpServlet {

  @Override
  protected void doPost(HttpServletRequest request, HttpServletResponse response)
      throws ServletException, IOException {
    LOGGER.info("Log out .....");
    HttpSession session = request.getSession();

    session.removeAttribute("loginInfo");
    session.invalidate();

    response.sendRedirect("/");
  }
}
